requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
sqlalchemy==2.0.43
pydantic==2.11.7
python-dotenv==1.1.1
streamlit==1.49.1
pandas==2.3.2
openpyxl==3.1.5
requests==2.32.5
